ENGR00223788 mx6sl_evk: change boot args to boot from mmc/sd
default setting is to boot via nfs. By this way, after mfgtool burned the whole
image on EVK, the system will probably not load the Linux kernel if the network
is not correctly set or not availabe for it. Change to mmc/sd boot to make
sure an entire boot process can be completed, in order to easily verify mfgtool.

Signed-off-by: Robby Cai <R63905@freescale.com>